Alterando Calculadora para utilizar double ao invés de int

+++ 2025-2/PDM1/BemCalcula/app/src/main/java/com/oszika/bemcalcula/util/Calculadora.java
@@ -14,19 +14,19 @@ public class Calculadora {
     public Calculadora() {
     }
 
-    public double somar(int num1, int num2) {
+    public double somar(double num1, double num2) {
         return num1 + num2;
     }
 
-    public double subtrair(int num1, int num2) {
+    public double subtrair(double num1, double num2) {
         return num1 - num2;
     }
 
-    public double multiplicar(int num1, int num2) {
+    public double multiplicar(double num1, double num2) {
         return num1 * num2;
     }
 
-    public double dividir(int num1, int num2) {
+    public double dividir(double num1, double num2) {
         if (num2 == 0) {
             throw new IllegalArgumentException(resources.getString(R.string.erro_divisao_por_zero));
         }
